Fix detection of EOCD block
Apparently, it's not actually a requirement of the ZIP-format that the
file ends after the EOCD block. I.e. a zip-file can contain additional
data after the EOCD block and comment, that is just ignored by other
ZIP-tools. Therefore, the test we do to make sure that the EOCD block
end coincides with the actual EOF breaks for this kind of ZIP file. This
commit removes the check, making http_zip more lenient when trying to
find the EOCD.

Closes #2